# backend/tests/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelChannelCreateIn": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "$ref": "#/components/schemas/OpenAICompatibleModelConfigIn"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "config"
        ],
        "title": "ModelChannelCreateIn",
        "type": "object"
      },
      "ModelChannelDraftUpdate": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "$ref": "#/components/schemas/OpenAICompatibleModelConfigIn"
          },
          "expected_revision_id": {
            "minimum": 1.0,
            "title": "Expected Revision Id",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision_id",
          "config"
        ],
        "title": "ModelChannelDraftUpdate",
        "type": "object"
      },
      "ModelChannelOut": {
        "additionalProperties": false,
        "properties": {
          "active_revision_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Revision Id"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "revisions": {
            "items": {
              "$ref": "#/components/schemas/ModelChannelRevisionOut"
            },
            "title": "Revisions",
            "type": "array"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "kind",
          "enabled",
          "active_revision_id",
          "created_at",
          "updated_at",
          "revisions"
        ],
        "title": "ModelChannelOut",
        "type": "object"
      },
      "ModelChannelRevisionOut": {
        "additionalProperties": false,
        "properties": {
          "base_host": {
            "title": "Base Host",
            "type": "string"
          },
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "secret_configured": {
            "title": "Secret Configured",
            "type": "boolean"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "tested_ok_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tested Ok At"
          }
        },
        "required": [
          "id",
          "state",
          "base_url",
          "base_host",
          "model",
          "secret_configured",
          "tested_ok_at",
          "created_at"
        ],
        "title": "ModelChannelRevisionOut",
        "type": "object"
      },
...
      "OpenAICompatibleModelConfigIn": {
        "additionalProperties": false,
        "properties": {
          "api_key": {
            "$ref": "#/components/schemas/SecretUpdate"
          },
          "base_url": {
            "maxLength": 2048,
            "minLength": 8,
            "title": "Base Url",
            "type": "string"
          },
          "kind": {
            "const": "OPENAI_COMPATIBLE",
            "default": "OPENAI_COMPATIBLE",
            "title": "Kind",
            "type": "string"
          },
          "model": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          }
        },
        "required": [
          "base_url",
          "model",
          "api_key"
        ],
        "title": "OpenAICompatibleModelConfigIn",
        "type": "object"
      },
...
      "SecretUpdate": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "enum": [
              "KEEP",
              "REPLACE",
              "CLEAR"
            ],
            "title": "Action",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "action"
        ],
        "title": "SecretUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkbenchUrlIn": {
        "additionalProperties": false,
        "properties": {
          "url": {
            "default": "",
            "maxLength": 2048,
            "title": "Url",
            "type": "string"
          }
        },
        "title": "WorkbenchUrlIn",
        "type": "object"
      },
      "WorkbenchUrlOut": {
        "additionalProperties": false,
        "properties": {
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "WorkbenchUrlOut",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Alert Workbench",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/model-channels": {
      "get": {
        "operationId": "list_model_channels_api_model_channels_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelChannelOut"
                  },
                  "title": "Response List Model Channels Api Model Channels Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Model Channels"
      },
      "post": {
        "operationId": "create_model_channel_endpoint_api_model_channels_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelChannelCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelChannelOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Model Channel Endpoint"
      }
    },
    "/api/model-channels/{channel_id}/disable": {
      "post": {
        "operationId": "disable_model_channel_endpoint_api_model_channels__channel_id__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "required": true,
            "schema": {
              "title": "Channel Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelChannelOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable Model Channel Endpoint"
      }
    },
    "/api/model-channels/{channel_id}/draft": {
      "put": {
        "operationId": "update_model_channel_endpoint_api_model_channels__channel_id__draft_put",
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "required": true,
            "schema": {
              "title": "Channel Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelChannelDraftUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelChannelOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Model Channel Endpoint"
      }
    },
    "/api/model-channels/{channel_id}/enable": {
      "post": {
        "operationId": "enable_model_channel_endpoint_api_model_channels__channel_id__enable_post",
        "parameters": [
          {
            "in": "path",
            "name": "channel_id",
            "required": true,
            "schema": {
              "title": "Channel Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelChannelOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enable Model Channel Endpoint"
      }
    },
...
    "/api/settings/workbench-url": {
      "get": {
        "operationId": "get_workbench_url_api_settings_workbench_url_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkbenchUrlOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Workbench Url"
      },
      "put": {
        "operationId": "put_workbench_url_api_settings_workbench_url_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkbenchUrlIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkbenchUrlOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Workbench Url"
      }
    }
  }
}